requirements.txt
contexttimer==0.3.3
python-dotenv==1.0.0
boto3==1.28.83
lxml==4.9.3
html5lib== 1.1
requests==2.31.0
pyarrow==14.0.1
polars==0.19.12
